Move confirm/input to UI components
